requirements.txt
# Core
python-telegram-bot==20.7
anthropic>=0.39.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Date/Time parsing
python-dateutil==2.8.2
pytz==2024.1

# HTTP and async
aiohttp==3.9.1
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.26.3

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
